DAOS-3810 api: container list drpc handler, more srv_drpc_tests (#1553)
With this change, a new daos_io_server drpc handler is implemented
to provide pool list containers functionality. This is a companion
to the existing C API and CaRT RPC infrastructure, and will be
engaged by the dmg control plane admin utility in a future commit.

Go and C unit tests for the container list functionality are provided.

Also the run_go_tests.sh script is updated to refer to the OFI
libraries and header file locations. Without these additions, test
build failures and subsequent omissions were observed in a developer
test environment.

Signed-off-by: Ken Cain <kenneth.c.cain@intel.com>