DAOS-3917 bio: Reduce monitor period after FAULTY state has occurred
Current monitoring period is set at 60 seconds, which is fine up until the
device state has changed to FAULTY. After this case, it will take too long in
most cases for the blobstore state transition from TEARDOWN->OUT because it has
to loop and check at a minimum a multiple of 60 sec(# of xstreams). Reduce this
monitor period to something more reasonable like 10 seconds for this case.

Also fix logic for state transition from FAULTY->TEARDOWN - blobstore can
transition to TEARDOWN if targets are in DOWN or DOWNOUT (not just DOWNOUT).

Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>