DAOS-2447 obj: refine EC handing for evenly distribution case
When the recxs evenly distributed to all targets for update, or all
data targets for fetch, need not regenerate new IOD/recxs or siods to
save RPC size.
It is helpful for singv handing as well as it is even distribution for
most cases.

Met a problem for checksum calculation, probably this patch is not very
helpful and make the code more complex. Upload just for backup now, can
pick it up later when needed.

Signed-off-by: Xuezhao Liu <xuezhao.liu@intel.com>