DAOS-3940 control: Move TableFormatter into txtfmt package
Enables reuse outside of dmg. Also cleans up the API to use
variadic parameters for titles so that callers don't need
to construct string slices.

Adds a new EntityFormatter type which can be used for neatly
displaying attributes of a single entity. The FormatEntity()
function can be used to do the formatting in a single operation.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>