DAOS-3960 DFS: add list-io API for read & write operation.
We need DFS read & write call that take an IOD with to be able to perform
list-io on a file. this is needed to switch the MPI-IO driver to use the
DFS API instead of the array API for IO to get other features from DFS
such as short reads and hole reading.

Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>