DAOS-623 build: Use mpicc when MPI_PKG is set if in path (#1668)
Updates scons_local and cart version so local build can use
custom MPI using MPI_PKG option for both daos and cart

If user sets MPI_PKG, it will first see if mpicc is in path and
use it to compile the MPI applications. Otherwise, it looks
for the specified pkg-config in PKG_CONFIG_PATH.

If MPI_PKG is not set, it will fall back to automatic search
using modules. We can probably remove this eventually
but it’s currently needed for CI

This allows for building with impi as it doesn’t provide pkgconfig
files. Essentially you can put mpicc in the path and run

scons MPI_PKG=any

to compile with it

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>