DAOS-3920 control: OFI_DOMAIN not set properly on some Mellanox configurations (#1604)
getNodeAlias improves the sibling match by ignoring siblings that are
found to be system devices.  This additional filter condition is helpful
when the fabric interface has more than one device associated with it.
For example, if the fabric adapter mlx4_0 has devices ib0 and enp2s0,
the sibling of ib0 will now be reported as mlx4_0 instead of enp2s0
because enp2s0 is a system device and mlx4_0 is not.  There may be
multiple system devices associated with a single fabric interface.
There will not be multiple fabric interfaces associated with one or more
of the same system devices.

The net result is that when the control plane automatically sets
OFI_DOMAIN, it will find the correct name to use.

Signed-off-by: Joel Rosenzweig <joel.b.rosenzweig@intel.com>