DAOS-2282 dfuse: Hook into dfuse in the container create/destroy command.
Add dfuse hooks in duns to allow container create to insert an entry
point in dfuse.
Fix a bug in setattr.
Return correct error codes if pool or container attach fail in dfuse.

This works correctly for creating a container, and the paths resolve
for container destroy however the destroy itself currently fails,
seemingly because the container is in use so further work is required here.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>