DAOS-4016 test: check return of pool_destroy_safe()
With this change the daos_test common function test_teardown()
now examines the return code of pool_destroy_safe() and frees the
test state in an error scenario. Before the change he return code
was not examined. In rare error scenarios daos_test instability was
observed following a failure to destroy a DAOS pool in teardown.

Signed-off-by: Ken Cain <kenneth.c.cain@intel.com>