DAOS-4027 pydaos: add dbm support (#1432)
Fix crashes in some kv operations where the type was not as
expected.
Change the output type to be non-truncated byte strings.
Change the install path in setup.py
Raise KeyError if key not found to match other dict like objects
Add support for del() of keys.
Add daosdbm for named KV stores.

Fix __contains__ in python kv stores. 
This would correctly return True if the key was found but would
raise an exception if it wasn't.  Catch this exception and return
False now.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Co-authored-by: Johann Lombardi <johann.lombardi@intel.com>