DAOS-3954 vos: not create punch target for non-rebuild case
For non-rebuild case, if the target (to be punched) does not exist,
it is not necessary to create the target before punch it, instead,
just return success directly; otherwise, it will leave useless
information in related tree and cause additional incarnation log
and DTX operations.

For DTX logic, if nothing is modified, then do not add it into CoS
cache, instead, release the DTX directly.

Signed-off-by: Fan Yong <fan.yong@intel.com>