Fixing cleanup code.
Using a set for the cleanup commands resulted in commands being out of
order for the removal of the superblock and umounting the scm_mount.

Also fixed th code that detects a failed 'daos_server start' to raise an
error if unable to detect the correct number of io server start mesages
after formatting.

Also moved hdf5.py to use dcpm/nvme to resolve DER_NOSPACE error.

Signed-off-by: Phillip Henderson <phillip.henderson@intel.com>