DAOS-2167 test: restore/simplify pool destroy flow in tests
With this change, temporary measures added in PR #1643 adding a pool
disconnect call before destroy(force=true) are restored to the original
form. This is possible due to recent fixes to DAOS pool destroy in
PR #1625. Also with this change, a centralized function to disconnect
from and remove pools used by a test specifies force=false when calling
the DAOS pool destroy API.

The pool bad_evict and bad_exclude test yaml configuration files
are updated so the agent and the daos_server configuration both use
the non-default DAOS system name (only daos_server used the name
causing the test to not run properly).

Several test yaml configuration files specify that the server should
be started with --recreate-superblocks, to adjust to changes made
by PR #1789 that requires tests to opt-in to this configuration.

Skip-run_test: true
Skip-func-hw-test: true
Test-tag: pr,-hw badevict badexclude container,full_regression rebuildcontcreate object,full_regression

Signed-off-by: Ken Cain <kenneth.c.cain@intel.com>