DAOS-4107 build: Add /usr/lib to rpath when PREFIX is specified (#66)
This is a temporary hack to get the docker build to work as
documented.   The real solution is to change all of our
component builds to install in lib64.  But that is a longer
term fix.   This adds /usr/lib to the rpath so that
dependencies of dependencies get picked up on the link
line without adding explicit link options.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>