DAOS-4107 build: Update scons_local for docker build fix (#1825)
When using PREFIX=/usr, we install prerequisite libraries
in /usr/lib rather than /usr/lib64.   This should be fixed.
Before that is fixed, in order to get the linker to pick
up dependencies without explicit linking, we must add
the non-standard link path to the rpath.  We only do
this if we build the component and its prefix is in
/usr

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>