DAOS-4132 iosrv: fix compiling issue with hwloc 2.x
In hwloc 2.x the allowed_cpuset is removed from hwloc_obj_t, and only
exports it by whole hwloc_topology and query it by
hwloc_topology_get_allowed_cpuset().

In DAOS usage, we don't set the HWLOC_TOPOLOGY_FLAG_WHOLE_SYSTEM for
hw_togology, so the allowed_cpuset is same as cpuset. So this patch just
replace the usage of allowed_cpuset by cpuset which is still exported.

Signed-off-by: Xuezhao Liu <xuezhao.liu@intel.com>