DAOS-4143 tests: Enable replacing bdev_list test yaml entries (#1884)
A '--nvme <value>' option has been added to launch.py to enbale
replacing NVMe device specifications in the functional tests it
executes.  Similar to the '--test_servers' and '--test_clients'
command line options, used to replace the hosts specified in the
test yaml file, this new option replaces any NVMe Bus:Device.Function
notations in the test yaml with each comma-separated BDF in the <value>
string, in corresponding order.

The special "auto[:<filter>]" keywoard can also be specifed as the
<value> string, which will result in the <value> string being assigned
to the BDF notations detected on the '--test_servers' hosts.  This
enables running the functional tests on clusters with different hardware
configurations - eliminating the need for changes to test yaml files
when hardware configurations are changed.

In addition, the DAOS-4069 bug has also been resolved by this change.

Signed-off-by: Phillip Henderson <phillip.henderson@intel.com>