DAOS-1781 api: check for invalid uuid in DAOS APIs
Several DAOS APIs accept a uuid_t parameter and perform a uuid_copy
involving that argument. A validity check is added on the argument
to protect against a segmentation fault that can occur in uuid_copy
if the argument is a NULL pointer.

A container create functional test is updated to no longer skip
negative test cases that supply a NULL pointer as the uuid
argument to container create.

Test-tag: pr,-hw containercreate

Signed-off-by: Ken Cain <kenneth.c.cain@intel.com>