DAOS-4100 container: integer overflow on sleep time
There is an integer overflow in cont aggregation code which leads
to the aggregation ULT being put to sleep for a very long time.

This patch also changed the criterion of dss_xstream_is_busy() to
"No IO requests in 5 seconds" instead of "No active requests in
this moment", because even if there is continuous IO flow from
client, the active request counter could be 0 at certain point,
that may result in the aggregation being triggered unexpectedly.

Signed-off-by: Niu Yawei <yawei.niu@intel.com>