CORCI-682 build: Publish on landing to $daos_branch
Which is defined at the top of the Jenkinsfile for each branch when a
branch  is created.  Would be better to figure this out dynamically
somehow but I don't know if that is possible.

Add Skip-test-centos-rpms pragma and remove the skip of testing CentOS
RPMs when Quick-build is true.

Don't skip checkpatch when quick-building.

Skip-checkpatch: true
Quick-build: true
Skip-test: true

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>