DAOS-4240 documentation: Refresh patch required to update docs
Skip-build: true
Skip-test: true

This patch contains changes to the Index.md, mkdocs.yaml and
terminology. Readme files that were refreshed are also captured.
A Dev section is added to capture the changes made to the
applicable /src/<component>/README.md files.
An initial version of the API documentation is included (doc/html)
as well as the doxyfile used to compile.
Mkdocs buils and mkdocs deploy must be completed for changes
to land on daos.io.

Signed-off-by: craig <craig.durfey@intel.com>