CORCI-835 build: Get quickbuild deps from caller (#1578)
Currently we have to maintain the proper versions of dependencies in the
Dockerfile.centos.7 file when used with Quick-build: true.  It gets out
of date often.

Instead, the list of packages to install for Quick-build: true should be
given to the Dockerfile by Jenkins after it queries the daos.spec to get
the list of dependencies.

Make RPM build log step more obvious.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>