DAOS-4244 iosrv: ignore ABT_thread_attr_free failure
The ABT_thread_attr_free() may fail after the child ULT created.
We should not return such failure to the caller. Otherwise, the
caller will free the parameters "arg" that is being used by the
child ULT as to segfault.

So let's ignore the failure, the worse case is that we may leak
some DRAM, but that is not fatal.

Signed-off-by: Fan Yong <fan.yong@intel.com>