DAOS-4270 control: Preserve server-side faults for clients
Allow clients to receive structured Fault errors in order to
display resolutions when available.

Also adds a fault indicating that the harness has not started
rather than the cryptic "not an access point" message. This
tends to happen when the ioservers haven't started for whatever
reason (waiting for format, config problems, etc).

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>