DAOS-4272 tests: Resolving D_LOG_FILE not being archived
The client log file used with functional tests like daos core tests,
ior, and mdtest was not being archived with the test results.  Now that
the TestWithServers log file attibutes are only names an additional
method is required to add the path to the client log file names such
that they will be archived by launch.py.

Signed-off-by: Phillip Henderson <phillip.henderson@intel.com>