DAOS-1946 md: fix initial solution for metadata full handling
The original solution for rdb_raft_state checking to handle
-DER_NOSPACE condition was to trigger log compaction and step down
as the service leader. A problem with this implementation in commit
a614cb1b64bf0855ea8c298b392d1149ea0cf72f is that it compacts to the
current index. With this fix, it compacts to the committed index.

Test-tag-hw-large: pr,hw,large metadatafill metadata_free_space

Signed-off-by: Ken Cain <kenneth.c.cain@intel.com>