DAOS-4246 srv: Update to lastest Cart
Let's remove module choose in rpc handler temporarily, which
is not used currently, because this callback will be shared
by rpc handler and rpc send callback as well, so the argument
might be rpc or cb_info.

Do not overwrite the failure in obj_rw_local().

Compare the remote bulk size with the size fetched to return
correct failure(E2BIG) for client to retry.

Signed-off-by: Di Wang <di.wang@intel.com>