DAOS-365 container: Mind handles in cont_destroy (#1949)
When destroying a container with existing handles, cont_destroy should
return -DER_BUSY if force == false or evict all existing handles if
force == true. To implement this, we add a handle index KVS under each
container property KVS, in order to quickly identify the set of all
handles for a container.

Signed-off-by: Li Wei <wei.g.li@intel.com>