DAOS-4267 control: Use system.Rank consistently
Renames ioserver.Rank -> system.Rank, as this fits better
and is more easily importable without baggage. Also changes
all uint32-based server ranks to system.Rank.

The intent behind this change is to simplify the logic surrounding
ranks. A rank should only be nil/unset under very specific
circumstances, so most uses should not involve pointers or
conversion between types. Using the type alias consistently throughout
the codebase makes it easier to read as well, in that the reader
sees that the value is a system.Rank without having to deduce based
on context.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>