DAOS-1965 md: Fix regression on metadata_add_remove test (#2114)
The maximum number of containers created before which metadata space
becomes exhausted has changed (lower) due to changes in DAOS both on
the master branch and release/0.9 branches. This has led to CI test
regressions. Lower the value of the constant in the test to 98% of
the lowest observed number of containers, factoring in differences
in behavior across the two branches.

Signed-off-by: Ken Cain <kenneth.c.cain@intel.com>