DAOS-3113 vos: Conditional punch/insert/update (#1989)
This patch adds conditional update, insert, and punch support to DAOS.

It includes the following changes:
To make some space for a timestamp index, the punch flag in the root entry of an incarnation log is now a single bit.
Read timestamps are implemented using a flat array as an LRU cache. The index into the array is stored in persistent memory and the virtual address of the index is used a key to detect cache misses. This flat array has positive and negative entries for each level of the tree. If an entry is evicted, timestamps are updated in the parent negative entry (or global time). This is not expected to be a common occurrence but should it happen, the result is false conflicts.
VOS API is updated to add a flags parameters, including a special flag (VOS_OF_USE_TIMESTAMPS) to differentiate between vos calls that need to check the timestamp cache and those that don't (such as rebuild).
The DAOS API is now hooked up to the VOS API. The flags in the DAOS API have been updated slightly. For instance, there is no need to differentiate DKEY punch and AKEY punch in the flags because it's implied by the API which entity is being punched.
Updates DFS test to fix a few bugs with conditional update
Updates a few unit tests to cut down on memory usage to avoid overloading our CI VMs. The 80MB timestamp cache caused swapping so we must have already been pretty close to the limits.
It adds vos_ilog_punch for common handling of punch logic since existence checks are required for conditional punch.
Add a simple unit test
Conditional tests are disabled in VMEM mode

Fan Yong added the following:
For conditinoal update case, the former committable update may
conflict with subsequent conditional update. Let's commit them
via CoS mechanism.

Mohamad made a few minor fixes to DFS conditional operations

Co-authored-by: Fan Yong <fan.yong@intel.com>
Co-authored-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>