Added Basic Tests for DMG Reintegrate.
Added some basic tests to pool_tests.go as well as
drpc tests.

There still needs to be some discussion on how to handle
input that has both valid and invalid values, such as
target idxs that are up as well as ones that are down.

This will probably move to a later ticket once solidified.

Signed-off-by: Peter Fetros <peter.fetros@intel.com>