DAOS-4295 dfuse: Check container and pool uuids better at start.
dfuse has to go into the background before calling daos_init() for
daos to work properly, so there are lots of failure modes at start
that cannot be returned properly to the user.  Checking that uuids
passed for pools and containers however, can be checked so they
should be, as that allows us to give higher quality feedback
when they are incorrect.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>