DAOS-4347 Agent: Autodetect client network interface
DRAFT

Adds DetectNetworkInterface() to netdetect which scans the
fabric to find the highest performance network interface
matching the given provider and NUMA node.

Uses DetectNetworkInterface inside the agent's handleGetAttachInfo()
and transmits the results back to the client in the rpc response.

This functionality depends on landing DAOS-2489 and DAOS-4029

Skip-build: true
Skip-test: true

Signed-off-by: Joel Rosenzweig <joel.b.rosenzweig@intel.com>