DAOS-2167 api: add a force destroy pool test (#2009)
With this change a new pool destroy test is added that invokes the
DAOS pool destroy API with force=1 on a connected pool. Before recent
changes to the pool destroy implementation, this scenario was observed
to result in segmentation faults. This does not occur with the latest
implementation. So now a test is needed to explicitly verify.

Signed-off-by: Ken Cain <kenneth.c.cain@intel.com>