DAOS-3717 control: Allow optional syslog logging
When the logging system was designed, the expectation was that
server management would be done via systemd, which expects
log output on stdout and adds it to the journal. In some
cases it may be desirable to log to /var/log/messages instead,
so supplying the --syslog flag to daos_server start will
accomplish this.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>