DAOS-4029 agent: Cache GetAttachInfo data on agent (#2146)
Stores data from the first GetAttachInfo response
to avoid making this RPC to the daos_server
in the future.  Subsequent requests for data
will return the cached elements to improve performance.

It is expected that the agent is torn down when
the cached data becomes stale. Therefore, there is 
no cache writeback policy or cache flush implemented
on the server side.

In order to disable the caching and revert to the previous
behavior where the agent always makes a request to the
control plane, set DAOS_AGENT_DISABLE_CACHE=true
in the agent's environment.

Signed-off-by: Joel Rosenzweig <joel.b.rosenzweig@intel.com>