DAOS-4336 Security: Unify certificate naming
The default values for certificates expected by each component do not line up
with the names that are generated by the certificate generation script.
Additionally dmg and daos_agent shared a single default configuration. This
splits those configurations to have component specific configuration defaults
in addition to unifying the default values of the certificate names.

Signed-off-by: David Quigley <david.quigley@intel.com>