DAOS-3862 dfs: fetch symlink value in a separate RPC (#2313)
Depending on the type of the entry being fetch, add a second fetch
to get the symlink value instead of doing it in the original fetch.
The original way was adding overhead to all entry types stats which
we can avoid by separating the symlin fetch because that needs a bulk.

Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>
Conflicts:
	src/tests/suite/dfs_test.c