DAOS-3905 dtx: cache non-committed modifications on client
For distributed transaction, before client triggering commit,
the modifications will be cached on client side. The benefit
is that if the distributed transaction needs to be restarted
because of (potential) conflict with others, we just need to
drop the cached modifications on client without talking with
related daos servers. On the other hand, if the distributed
transaction owner (the application process or client) crashed
before committing the distributed transaction, server needs
to nothing to cleanup related distributed transaction.

But caching modifications on client is not free, especially
consider the case that the application may wants to modify
a lot of data via distributed transaction. Under such case,
if we copy the application buffers into the cache, it will
not only waste a lot of CPU cycles, but also consume client
DRAM very much. To avoid such trouble, the DAOS API (update
and punch) allows the application to indicate the DTX logic
whether needs to copy application buffer or not. If specify
the 'DAOS_ZERO_COPY' flag explicitly when call the DAOS API,
then the DTX logic will directly reference related pointers
to the application buffers when cache. Under such case, the
application needs to guarantee that do not reuse or release
related buffers until the distributed transaction committed
(by client); otherwise, the non-committed modification will
be overwritten or lost.

Signed-off-by: Fan Yong <fan.yong@intel.com>