DAOS-4633 vos: Propagate punches when result is empty tree
In order to avoid emptiness checks on iteration, we can instead
propagate punches that result in an empty tree to the parent
key/object.   Thus we pay a penalty on punch but make it up
on update.

Adds an internal iterate function based on an open tree handle
to do the emptiness check.
Modifies punch so that if the object/dkey/akey are empty, it
is a noop unless it's a replay punch.  This matches the behavior
prior to adding the incarnation log.   Such punches will be
aggregated anyway.

Modify mvcc tests so that they fail fast if CMOCKA_TEST_ABORT
is set.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>