Ensuring the ior command uses the correct test file argument.
When multiple IOR commands are issued through the same IorCommand object
and one of the previous commands utilizes the POSIX API the
IorCommand.test_file parameter is modified, but never returned back to
its previous value.  This change ensures that the test_file value is
restored when not using the POSIX API for the same IorCommand object.

Signed-off-by Phillip Henderson <phillip.henderson@intel.com>