DAOS-4451 md: temporary logging for container/create failures
Running soak stress testing of RPMs created from the release/0.9 branch
revealed (in a 48 hour period) two instances of container create
failures. Error logging is kept to a minimum for long stress test
runs like this. With this change, temporarily add D_ERROR logging to
the container create code paths to capture some preliminary information
on the nature of the failures.

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>