DAOS-623 vos: Fix VOS size check for integer keys
The feats were not being passed to the overheads check.
VOS integer and lexical keys both use direct key.  It
occurs to me that we could change this for integer keys
now because we no longer need both a key and an epoch
since the incarnation log was added but that is for
another day.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>