DAOS-4532 vos: Use built-in btree integer keys for vos
Before the incarnation log, we required the epoch in the
key tree as part of the comparison.   Now we can use the
integer key directly and save some overhead.

Rather than dereferencing the record and calling vos to
do comparisons, we can just store the integer key directly
in the tree.

This is a format change but seems like a good thing to do.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>