DAOS-4158 csum: Align chunk size to record size (#2351)
Force checksum chunk sizes to be a multiple of the record size. If the
record size is larger than the chunk size, set the chunk size to be
the record size. Will never fetch partial records so it wouldn't make
sense to try to create checksums for part of a record, even if the
record is larger than the configured chunk size.

Signed-off-by: Ryon Jensen <ryon.jensen@intel.com>