DAOS-4263: btree_direct cleanup
For btree_direct, this commit implements the execution of a sequence
of tests in a single cmocka instance, instead of a cmocka instance
per argument.

For this, a --start-test prameter is used to define the name of the
cmocka exetuion. If this parameter is not provided, then a default
test name is used.

Signed-off-by: Marcela Rosales <marcela.a.rosales.jimenez@intel.com>