DAOS-4679 test: Reset runable bool for test_runable() check
When one check of test_runable() for daos_test returns false (ie not enough
servers to run test, etc), then the value is not reset to true at the beginning
of each new test case check. Instead the runable bool is just initially set to
true in the static declaration.

Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>