DAOS-4488 dtx: handle conditional modification via DTX
For distributed transaction, we caches modifications on client
before being committed. Single conditional modification is equal
to atomic "fetch" + "update/punch". If such conditional operation
is inside distributed transaction, if we still cache non-committed
modifications, then we will not handle "fetch" that is the first
part of the conditional modification. So the result for the left
part modification may be different from the case of being executed
on the server. That may further misguide the application to make
take more wrong actions in subsequent operations.

The solution is that:

The DTX client side logic will interpret conditional modification
if it is inside DTX context, trigger conditional read (query) RPC
for related dkey/akey to related server, according to conditional
read result to decide whether cache the left modification part on
client or return error to application.

Signed-off-by: Fan Yong <fan.yong@intel.com>