DAOS-4667 vos: Use 16-bit local identifier and epoch in VOS
Replaces the umem_off_t in VOS trees with a 16-bit identifier
The ID is used in conjunction with the epoch to uniquely
identify the transaction in VOS trees.
The allocation uses the LRU array and can support up to 32K
inflight transactions at a time (this number could actually
be increased to 64K - 2 since we really only reserve two values
for committed and aborted entries.

Also, vos_update_begin and vos_obj_punch are modified to use
the dth epoch if it exists rather than the passed in one.  I
believe this only differs for VOS tests.

The data structures are still all the same size in the tree
to avoid too much change in one patch.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>