DAOS-4347 Agent: Autodetect client network interface (#2359)
The agent now automatically determines a network interface for the client.
When the client issues a GetAttachInfo dRPC, it invokes the Agent's
handleGetAttachInfo() which issues the gRPC to the daos_server. The
response contains the server's network provider. The agent scans the
local network interfaces to find devices that match the NUMA affinity of the
client application and support the server's provider.

There may be more than one compatible network interface on a given
NUMA node. The agent load balances between these interfaces using a
round-robin selection scheme. If the client is bound to a NUMA node with
no compatible devices, an error is returned.

The agent returns the response to the client containing an interface and
related domain that should be used when initializing CaRT via daos_init().

A future PR for DAOS-4348 will call this dRPC prior to daos_init() in order
to complete the client configuration. Manual interface configuration is
required until then.

Signed-off-by: Joel Rosenzweig <joel.b.rosenzweig@intel.com>