DAOS-4624 test: Test daos pool query and attribute
Query test: Create a pool and call daos pool query. From the output,
verify the UUID matches the one that was returned when creating the
pool. Also verify the pool size against the value that's passed in to
dmg pool create.

Attribute test:
1. Create 5 attributes to a pool. We can set attribute
name and corresponding value, so set unique name-value pair for each of
the 5 attributes.
2. Call pool list-attr. This returns the 5 attribute names, so compare
them against the actual names used.
3. Call pool get-attr. This returns the value of the corresponding
name, so do this for each of the 5 attributes.

Test-tag: pr,-hw pool_query_attr
Signed-off-by: Makito Kano <makito.kano@intel.com>