DAOS-3978 control: system reformat after controlled shutdown
Add 'dmg system format --system' to enable reformat of storage and metadata
after a controlled shutdown. Rank identifiers will be reset and
currently the only reliable use case will be to reformat all after
ranks in a system for the moment (drain and reintegration still need
to be integrated into the control-plane "system" command flows).

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>