DAOS-4653 control: Update control configuration
As part of the move to the new Control API, the configuration
file has been renamed from daos.yml to daos_control.yml in order
to better convey its relationship to what's being configured.

In addition to this renaming and general code reorganization,
a check has been added to exit with an informative error when
a hostlist is supplied for a dmg command that connects to the
Management Service (e.g. pool operations). This is necessary
to ensure that the correct AP is chosen when the Replicated
Management Service feature lands.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>