DAOS-4423 cart: Fix gp_self initialization
For a secondary group (SG) CoRPC sent by SG rank 0, another participant
may mistakenly think itself is the root, completing this CoRPC without
sending a reply to its parent. This was observed in a system of a single
server node with 2 servers (by printing the variables in
crt_corpc_complete):

  crt_corpc_complete(0x7f394c1351f0)  [opc=0xff00000e xid=0x1
  rank:tag=1:0] am_root=1 myrank=0 root=0 co_info=0x7f394c135e50

The root cause is that gp_self is initialized to 0 for newly-created
SGs. The CoRPC arrived before gp_self was set to a valid rank (by the
latest SG version), confusing crt_corpc_complete.

This patch initializes gp_self to CRT_NO_RANK for newly-created SGs and
group views. An effort is also made to make sure that gp_self is checked
against CRT_NO_RANK before being used. If an operation finds gp_self to
be CRT_NO_RANK, it is generally because the group hasn't received the
latest membership yet. Therefore, we return -DER_GRPVER in these cases.

Signed-off-by: Li Wei <wei.g.li@intel.com>